Backend/job_board_api/job_board_api/views.py
from rest_framework import generics
from .models import JobPosting
from .serializers import JobPostingSerializer

class JobList(generics.ListCreateAPIView):
    queryset = JobPosting.objects.all()
    serializer_class = JobPostingSerializer

class JobDetail(generics.RetrieveUpdateDestroyAPIView):
    queryset = JobPosting.objects.all()
    serializer_class = JobPostingSerializer
