Backend/job_board_api/job_board_api/requirements.txt
asgiref==3.6.0
distlib==0.3.6
Django==4.1.12
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.4.2
Faker==19.12.0
filelock==3.12.0
pbr==5.11.1
platformdirs==3.3.0
pymongo==4.5.0
python-dateutil==2.8.2
pytz==2023.3.post1
six==1.16.0
sqlparse==0.2.4
stevedore==5.1.0
virtualenv==20.22.0
virtualenv-clone==0.5.7
virtualenvwrapper==4.8.4
